### 2021-11-29 #3959 @whedon
[REVIEW]: Dune-MMesh: The Dune Grid Module for Moving Interfaces
**Submitting author:** @samuelburbulla (<a href="http://orcid.org/0000-0002-2566-9777">Samuel Burbulla</a>)
**Repository:** https://github.com/samuelburbulla/dune-mmesh/
**Branch with paper.md** (empty if default branch): feature/paper
**Version:** v1.3.2

@krober10nd & @ @vijaysm, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] **Contribution and authorship:** Has the submitting author (@samuel.burbulla) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

## Review checklist for @ @vijaysm

- [x] **Contribution and authorship:** Has the submitting author (@samuel.burbulla) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
