### 2021-12-10 b888215
Reimplement interfaceindicator and normals in separate modules.
