### 2021-12-13 samuelburbulla/dune-mmesh#2 @krober10nd
Installation Error 
On a 2020 Macbook M1, I am unable to install ddune-mesh. It appears to be related to the installation of DUNE, which requires `openmp` by default.  Is it possible to compile this package with openmp? Is this listed as a required dependency?

Generally speaking if an application depends on an upstream dependency that requires many packages, these upstream packages should also be listed in the dependency list.

For example this:

```
 pip install dune-mmesh
```

produces (the truncated output)

```
    -- Build files have been written to: /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build
    [1/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/space/shapefunctionset/orthonormal/orthonormalbase_1d.cc.o
    [2/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/storage/singleton.cc.o
    FAILED: lib/CMakeFiles/dunefem.dir/__/dune/fem/storage/singleton.cc.o
    /Library/Developer/CommandLineTools/usr/bin/c++ -DENABLE_GMP=1 -DENABLE_MPI=1 -DHAVE_CONFIG_H -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6 -I/opt/homebrew/Cellar/python@3.9/3.9.9/Frameworks/Python.framework/Versions/3.9/include/python3.9 -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/include -I/opt/homebrew/include/eigen3 -isystem /opt/homebrew/Cellar/open-mpi/4.1.1_2/include -isystem /opt/homebrew/Cellar/gmp/6.2.1_1/include -std=c++17  -Xclang -fopenmp -O3 -DNDEBUG -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX12.0.sdk -mmacosx-version-min=12.0 -fPIC -MD -MT lib/CMakeFiles/dunefem.dir/__/dune/fem/storage/singleton.cc.o -MF lib/CMakeFiles/dunefem.dir/__/dune/fem/storage/singleton.cc.o.d -o lib/CMakeFiles/dunefem.dir/__/dune/fem/storage/singleton.cc.o -c /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/storage/singleton.cc
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/storage/singleton.cc:3:
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/storage/singleton.hh:13:
    /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/threads/threadmanager.hh:26:10: fatal error: 'omp.h' file not found
    #include <omp.h>
             ^~~~~~~
    1 error generated.
    [3/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/space/shapefunctionset/legendrepolynomials.cc.o
    [4/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/solver/rungekutta/butchertable.cc.o
    [5/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/quadrature/pardgsimplexquadrature0d.cc.o
    [6/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/io/io.cc.o
    [7/19] Building CXX object python/dune/fem/CMakeFiles/_fem.dir/_fem.cc.o
    FAILED: python/dune/fem/CMakeFiles/_fem.dir/_fem.cc.o
    /Library/Developer/CommandLineTools/usr/bin/c++ -DENABLE_GMP=1 -DENABLE_MPI=1 -DHAVE_CONFIG_H -D_fem_EXPORTS -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6 -I/opt/homebrew/Cellar/python@3.9/3.9.9/Frameworks/Python.framework/Versions/3.9/include/python3.9 -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/include -I/opt/homebrew/include/eigen3 -isystem /opt/homebrew/Cellar/open-mpi/4.1.1_2/include -isystem /opt/homebrew/Cellar/gmp/6.2.1_1/include -std=c++17  -Xclang -fopenmp -O3 -DNDEBUG -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX12.0.sdk -mmacosx-version-min=12.0 -fPIC -MD -MT python/dune/fem/CMakeFiles/_fem.dir/_fem.cc.o -MF python/dune/fem/CMakeFiles/_fem.dir/_fem.cc.o.d -o python/dune/fem/CMakeFiles/_fem.dir/_fem.cc.o -c /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/python/dune/fem/_fem.cc
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/python/dune/fem/_fem.cc:6:
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/mpimanager.hh:9:
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/quadrature/caching/registry.hh:13:
    /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/threads/threadmanager.hh:26:10: fatal error: 'omp.h' file not found
    #include <omp.h>
             ^~~~~~~
    1 error generated.
    [8/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/space/shapefunctionset/orthonormal/orthonormalbase_3d.cc.o
    [9/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/misc/femtimer.cc.o
    FAILED: lib/CMakeFiles/dunefem.dir/__/dune/fem/misc/femtimer.cc.o
    /Library/Developer/CommandLineTools/usr/bin/c++ -DENABLE_GMP=1 -DENABLE_MPI=1 -DHAVE_CONFIG_H -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6 -I/opt/homebrew/Cellar/python@3.9/3.9.9/Frameworks/Python.framework/Versions/3.9/include/python3.9 -I/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/include -I/opt/homebrew/include/eigen3 -isystem /opt/homebrew/Cellar/open-mpi/4.1.1_2/include -isystem /opt/homebrew/Cellar/gmp/6.2.1_1/include -std=c++17  -Xclang -fopenmp -O3 -DNDEBUG -arch arm64 -isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX12.0.sdk -mmacosx-version-min=12.0 -fPIC -MD -MT lib/CMakeFiles/dunefem.dir/__/dune/fem/misc/femtimer.cc.o -MF lib/CMakeFiles/dunefem.dir/__/dune/fem/misc/femtimer.cc.o.d -o lib/CMakeFiles/dunefem.dir/__/dune/fem/misc/femtimer.cc.o -c /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/femtimer.cc
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/femtimer.cc:3:
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/mpimanager.hh:9:
    In file included from /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/quadrature/caching/registry.hh:13:
    /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/dune/fem/misc/threads/threadmanager.hh:26:10: fatal error: 'omp.h' file not found
    #include <omp.h>
             ^~~~~~~
    clang: error: unable to execute command: Segmentation fault: 11
    clang: error: clang frontend command failed due to signal (use -v to see invocation)
    Apple clang version 13.0.0 (clang-1300.0.29.3)
    Target: arm64-apple-darwin21.1.0
    Thread model: posix
    InstalledDir: /Library/Developer/CommandLineTools/usr/bin
    clang: note: diagnostic msg: Error generating preprocessed source(s).
    [10/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/space/shapefunctionset/orthonormal/orthonormalbase_2d.cc.o
    [11/19] Building CXX object lib/CMakeFiles/dunefem.dir/__/dune/fem/space/lagrange/shapefunctionset.cc.o
    ninja: build stopped: subcommand failed.
      File "/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/lib/python3.9/site-packages/skbuild/setuptools_wrap.py", line 593, in setup
        cmkr.make(make_args, env=env)
      File "/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/lib/python3.9/site-packages/skbuild/cmaker.py", line 504, in make
        raise SKBuildError(

    -- Trying "Ninja" generator
    --------------------------------
    ---------------------------
    ----------------------
    -----------------
    ------------
    -------
    --
    --
    -------
    ------------
    -----------------
    ----------------------
    ---------------------------
    --------------------------------
    -- Trying "Ninja" generator - success

    Configuring Project
      Working directory:
        /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build
      Command:
        cmake /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6 -G Ninja -DCMAKE_INSTALL_PREFIX:PATH=/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-install -DPYTHON_EXECUTABLE:FILEPATH=/Users/keithroberts/codes/dune-env/bin/python3.9 -DPYTHON_VERSION_STRING:STRING=3.9.9 -DPYTHON_INCLUDE_DIR:PATH=/opt/homebrew/opt/python@3.9/Frameworks/Python.framework/Versions/3.9/include/python3.9 -DPYTHON_LIBRARY:FILEPATH=/opt/homebrew/opt/python@3.9/Frameworks/Python.framework/Versions/3.9/lib/libpython3.9.dylib -DSKBUILD:INTERNAL=TRUE -DCMAKE_MODULE_PATH:PATH=/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-tk7tsuml/overlay/lib/python3.9/site-packages/skbuild/resources/cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_RPATH_USE_LINK_PATH=TRUE -DDUNE_ENABLE_PYTHONBINDINGS=TRUE -DDUNE_PYTHON_INSTALL_LOCATION=none -DALLOW_CXXFLAGS_OVERWRITE=ON -DCMAKE_DISABLE_FIND_PACKAGE_LATEX=TRUE -DCMAKE_DISABLE_FIND_PACKAGE_Doxygen=TRUE -DINKSCAPE=FALSE -DCMAKE_BUILD_TYPE:STRING=Release -DCMAKE_OSX_DEPLOYMENT_TARGET:STRING=12.0 -DCMAKE_OSX_ARCHITECTURES:STRING=arm64

    Traceback (most recent call last):

    An error occurred while building with CMake.
      Command:
        cmake --build . --target install --config Release --
      Source directory:
        /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6
      Working directory:
        /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-nhjm7qqu/dune-fem_502b2276c71b4df08e888c6e8473f9e6/_skbuild/macosx-12.0-arm64-3.9/cmake-build
    Please see CMake's output for more information.
    ERROR: Failed building wheel for dune-fem
  Failed to build dune-fem
  ERROR: Could not build wheels for dune-fem, which is required to install pyproject.toml-based projects
WARNING: Discarding https://files.pythonhosted.org/packages/6b/35/8589bdd45b0bd81c191df60882d4c6b57c5cd4a72db4c1a7c59dcafce97b/dune-mmesh-1.3.0.tar.gz#sha256=d851f20949f2461bba4ee4ae0bfc382b3ed58b8730f9517f5b4ec1706c4c87c6 (from https://pypi.org/simple/dune-mmesh/) (requires-python:>=3.4). Command errored out with exit status 1: /Users/keithroberts/codes/dune-env/bin/python3.9 /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-standalone-pip-zmy2yn_v/__env_pip__.zip/pip install --ignore-installed --no-user --prefix /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-680ia38n/overlay --no-warn-script-location --no-binary :none: --only-binary :none: -i https://pypi.org/simple -- setuptools wheel scikit-build cmake ninja requests 'dune-fem<=2.8.0.0' gmsh 'dune-grid<=2.8.0.0' Check the logs for full command output.
ERROR: Could not find a version that satisfies the requirement dune-mmesh (from versions: 1.2.0.dev20210413, 1.2.0.dev20210715, 1.2.0.dev20210825, 1.2.0.dev20210826, 1.3.0.dev20210831, 1.3.0.dev20210910, 1.3.0.dev20210924, 1.3.0.dev20211012, 1.3.0, 1.3.1)
ERROR: No matching distribution found for dune-mmesh
```
