### 2021-12-14 a8827b1
Fix interfaceIndicator for filteredGrid.
