### 2021-12-14 samuelburbulla/dune-mmesh#2 @samuelburbulla
OpenMP is considered as an optional dependency in dune-fem. All includes are wrapped by an `#ifdef _OPENMP`, i.e. we shouldn't get an error if OpenMP is not present.

On the M1 Mac, this variable is somehow defined, and yet we are not able to find the corresponding header.

I solved this issue on my M1 Mac by installing `libomp` via brew and setting some CXX flags:

1. `brew install libomp`
2. `export CXXFLAGS="-lomp -L/opt/homebrew/Cellar/libomp/13.0.0/lib/"`

Further, I observed other issues on the M1 that are not directly related to dune-mmesh or DUNE:
 - `scipy` has problems to be built with PEP518
 - `gmsh` has a problem with dynamic linking

So, within the virtualenv, to install scipy, I run

1. `pip install pybind11 cython pythran`
2. `pip install --no-binary :all: --no-use-pep517 scipy`

before installing dune-fem (which has scipy as dependency).

For gmsh we can run `pip install gmsh` (which gives 4.8.4), but we have to replace the `libgmsh.dylib` by one built locally from source of Gmsh.

1. Download _https://gmsh.info/src/gmsh-4.8.4-source.tgz_ and extract.
2. Step into, `mkdir build`, `cd build` `cmake ..`, `make`
3. `cp libgmsh.4.8.4.dylib YOUR-VENV/lib/python3.9/site-packages/gmsh-4.8.4-MacOSX-sdk/lib/libgmsh.dylib`

Glad to hear if this works for you.
