### 2021-12-15 samuelburbulla/dune-mmesh#2 @krober10nd
Thank you for the help. I was able to install it by exporting CXXFLAGS and pointing them to the location of the openmp header.

```
export CXXFLAGS="-I/opt/homebrew/Cellar/libomp/13.0.0/include -I/opt/homebrew/Cellar/boost/1.76.0/include/"
```

```
  Building wheel for dune-mmesh (pyproject.toml) ... done
  Created wheel for dune-mmesh: filename=dune_mmesh-1.3.1-cp39-cp39-macosx_12_0_arm64.whl size=5874334 sha256=0735f746960efdae1bca9e8571fb53ccc7321461559a6b9806a2f353fca1773d
  Stored in directory: /Users/keithroberts/Library/Caches/pip/wheels/2d/e6/e4/9e836d42cb9e21bb7e58976a5d297140f695e7901aaeb236db
Successfully built dune-mmesh
Installing collected packages: wheel, portalocker, mpi4py, dune-common, dune-geometry, dune-grid, fenics-ufl, dune-localfunctions, dune-istl, dune-alugrid, gmsh, dune-fem, dune-mmesh
Successfully installed dune-alugrid-2.8.0.post1 dune-common-2.8.0.post1 dune-fem-2.8.0.5 dune-geometry-2.8.0.post1 dune-grid-2.8.0.post1 dune-istl-2.8.0.post1 dune-localfunctions-2.8.0.post1 dune-mmesh-1.3.1 fenics-ufl-2019.1.0 gmsh-4.8.4 mpi4py-3.1.3 portalocker-2.3.2 wheel-0.37.0
```
I used the default Clang compilers.

I did not have to install the other packages "by hand" so to speak.

Perhaps this information can be included on the readthedocs website? I think it would be helpful.

However, now, how do I test this before jumping to the examples?
