### 2021-12-16 samuelburbulla/dune-mmesh#2 @samuelburbulla
Good to hear that it was that easy to fix the installation. I will add a comment to the RTD page.

Interesting that you had no problems with scipy and gmsh, I have to check what's going wrong on my M1.

We provide a test script in `dune-mmesh/scripts/test.py` that runs through all essential python functionality.
Unfortunately, it is not shipped with the packages on PyPI, but you can just download it from the repository (together with the `horizontal.msh` mesh file).
