### 2021-12-16 cfda74c
Add comment on OpenMP headers.
