### 2021-12-17 samuelburbulla/dune-mmesh#2 @krober10nd
Hm, something isn't right still. It appears to have been installed seemingly correctly but I'm not able to import modules.

```
(dune-env) keithroberts@Keiths-MacBook-Pro-2 dune-mmesh % pip list
Package             Version
------------------- -----------
beniget             0.4.1
Cython              0.29.25
dune-alugrid        2.8.0.post1
dune-common         2.8.0.post1
dune-fem            2.8.0.5
dune-geometry       2.8.0.post1
dune-grid           2.8.0.post1
dune-istl           2.8.0.post1
dune-localfunctions 2.8.0.post1
dune-mmesh          1.3.1
fenics-ufl          2019.1.0
gast                0.5.3
gmsh                4.8.4
mpi4py              3.1.3
numpy               1.21.4
pip                 21.3.1
ply                 3.11
portalocker         2.3.2
pybind11            2.8.1
pythran             0.11.0
setuptools          59.0.1
```

but

```
>>> from dune.grid import reader
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'dune.grid'
>>>
```
