### 2021-12-17 samuelburbulla/dune-mmesh#2 @krober10nd
Further, it's problematic having a dash in all the package names as Python doesn't recognize this in any import statement and it isn't clear how it should be written either.

Try to use this to solve the import issue doesn't help either.

https://stackoverflow.com/questions/8350853/how-to-import-module-when-module-name-has-a-dash-or-hyphen-in-it
