### 2021-12-22 samuelburbulla/dune-mmesh#2 @krober10nd
> Maybe the installation just failed because of a previous aborted installation.
Have you installed the packages after exporting the CXXFLAGS in a fresh virtual environment?
If not, could you try to install again in a fresh virtual environment?

I created a new virtual enviroment .

Executed this
```
pip install dune-mmesh
```

It says
```
(dune-env2) keithroberts@Keiths-MacBook-Pro-2 codes % pip install dune-mmesh
Collecting dune-mmesh
  Using cached dune_mmesh-1.3.1-cp39-cp39-macosx_12_0_arm64.whl
Collecting dune-grid<=2.8.0.5
  Using cached dune_grid-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting gmsh
  Using cached gmsh-4.8.4-py3-none-any.whl
Collecting dune-fem<=2.8.0.5
  Using cached dune_fem-2.8.0.5-cp39-cp39-macosx_12_0_arm64.whl
Collecting dune-localfunctions<=v2.8.0.5
  Using cached dune_localfunctions-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting dune-istl<=v2.8.0.5
  Using cached dune_istl-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting fenics-ufl==2019.1.0
  Using cached fenics_ufl-2019.1.0-py3-none-any.whl (282 kB)
Collecting dune-alugrid<=v2.8.0.5
  Using cached dune_alugrid-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting numpy
  Downloading numpy-1.21.5-cp39-cp39-macosx_11_0_arm64.whl (12.4 MB)
     |████████████████████████████████| 12.4 MB 5.4 MB/s
Collecting dune-geometry<=v2.8.0-r1
  Using cached dune_geometry-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting dune-common<=v2.8.0.post1
  Using cached dune_common-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
Collecting portalocker
  Using cached portalocker-2.3.2-py2.py3-none-any.whl (15 kB)
Collecting mpi4py
  Using cached mpi4py-3.1.3-cp39-cp39-macosx_12_0_arm64.whl
Collecting wheel
  Using cached wheel-0.37.0-py2.py3-none-any.whl (35 kB)
Installing collected packages: wheel, portalocker, numpy, mpi4py, dune-common, dune-geometry, dune-grid, fenics-ufl, dune-localfunctions, dune-istl, dune-alugrid, gmsh, dune-fem, dune-mmesh
Successfully installed dune-alugrid-2.8.0.post1 dune-common-2.8.0.post1 dune-fem-2.8.0.5 dune-geometry-2.8.0.post1 dune-grid-2.8.0.post1 dune-istl-2.8.0.post1 dune-localfunctions-2.8.0.post1 dune-mmesh-1.3.1 fenics-ufl-2019.1.0 gmsh-4.8.4 mpi4py-3.1.3 numpy-1.21.5 portalocker-2.3.2 wheel-0.37.0
```

Seems to install but then when you try to run any of the test or examples nothing is found.

```
>>> import dune
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'dune'
>>>

```

This is what `pip list` shows

```
(dune-env2) keithroberts@Keiths-MacBook-Pro-2 codes % pip list
Package             Version
------------------- -----------
dune-alugrid        2.8.0.post1
dune-common         2.8.0.post1
dune-fem            2.8.0.5
dune-geometry       2.8.0.post1
dune-grid           2.8.0.post1
dune-istl           2.8.0.post1
dune-localfunctions 2.8.0.post1
dune-mmesh          1.3.1
fenics-ufl          2019.1.0
gmsh                4.8.4
mpi4py              3.1.3
numpy               1.21.5
pip                 21.3.1
portalocker         2.3.2
setuptools          59.0.1
wheel               0.37.0
(dune-env2) keithroberts@Keiths-MacBook-Pro-2 codes %
```

Unfortunately I'm not able to go much further than this..

Okay about the name. But for the future, I would recommend following the PEP8 style which is pretty clear about the naming of packages...

http://www.python.org/dev/peps/pep-0008/
