### 2022-01-02 #3959 @krober10nd
Hey @diehlpk im on a holiday until the 6 of January then will continue where we left off. I tried a number of ways to install the package to no avail on my personal laptop. I can try with another OS and see if I have better luck. I’ve posted issues on the repository.
