### 2022-01-09 #3959 @krober10nd
@samuel.burbulla I have tried on a different clean laptop and OS and encountered the same error as I noted in your repo. The program appears to install but I cannot import or run any of the examples. I look forward to @vijaysm for help here.

I cannot procced further with the review of this software if I cannot install correctly...
