### 2022-01-10 samuelburbulla/dune-mmesh#2 @samuelburbulla
This is really weird and surprises me a lot. I apologise very much for the inconvenience…

Thanks for testing on another OS! Now we can be sure that a lot of things aren’t the source of error. Did you use the same versions of Python, Pip etc., which ones? I would really like to reproduce the error on my machine to fix this issue.

Some further ideas to try which will give us more info:

1. Is there something conspicuous when calling pip with verbose flag 'pip install -v dune-mmesh'? Could you provide me with the full log (maybe via mail)? I know that it is very long, but it contains a lot of helpful information about the build process.
2. Could you shortly try if version 1.3.0 installs correctly? There could have been some interfering changes in the backend of the dependent dune-common module. Just to make sure.
3. Could you try (or: Have you tried) to install the source version following the steps in https://dune-mmesh.readthedocs.io/en/latest/installation.html#from-source? This should also give us more verbose information about what is going wrong.

Thanks for you help to find the issue!
