### 2022-01-12 #3959 @vijaysm
@krober10nd My installation tries on my macbook have not been successful. The errors are related to OpenMP includes and I think you have already raised this issue. I haven't had the time to fully debug the issue.

I am also going to try on either a Fedora or an Ubuntu machine this week. I'll let you know how the installation proceeds. @samuel.burbulla
