### 2022-01-21 #3959 @samuelburbulla
Hi @vijaysm,
as you mentioned, we discussed this issue in [#2](https://github.com/samuelburbulla/dune-mmesh/issues/2) and I proposed the following solution:

````
1. brew install libomp
2. export CXXFLAGS="-lomp -L/opt/homebrew/Cellar/libomp/13.0.0/lib/"
````

Somehow, CMake finds the package OpenMP and sets OPENMP_FOUND, but afterwards the compiler does not find the `<omp.h>` header. I just cross checked this behaviour on my Mac both with and without OpenMP, but unfortunately I am not able to reproduce the issue.
