+++ pyproject.toml
@@ -1,3 +1,3 @@
 [build-system]
-requires = ['setuptools', 'wheel', 'scikit-build', 'cmake', 'ninja', 'requests', 'dune-fem<=2.8.0.5', 'gmsh', 'dune-grid<=2.8.0.5']
+requires = ['setuptools', 'wheel', 'scikit-build', 'cmake>=3.13', 'ninja', 'requests', 'dune-fem<=2.8.0.6', 'gmsh', 'dune-grid<=2.8.0.post1']
 build-backend = 'setuptools.build_meta'
