### 2022-01-26 #3959 @samuelburbulla
I created a notebook on Google Colab where I was able to install dune-mmesh and run a short example:

https://colab.research.google.com/drive/1CB8HohQ0gE2LU0PET-N2bTnElzRILBbR?usp=sharing

Maybe this is useful for testing code? At least, you might compare your setup with this one.
Unfortunately, the usability there is a bit unhandy in my opinion, and it takes its time to compile everything when starting a new runner.

Remark that in the meantime I uploaded a new package 1.3.1.post1 where I updated the version requirement of dune-fem to 2.8.0.6. Compared to version 2.8.0.5 some things regarding OMP have been fixed, maybe this already fixes your errors with omp.
