### 2022-01-27 4fd7ba9
Fix doc of installing dev version.
