### 2022-01-28 f1f5ec1
Add tuplespace adaptation test.
