+++ dune/mmesh/test/test-adapt-tuplespace.cc
@@ -1,0 +1,271 @@
+#include <config.h>
+
+#include <iostream>
+
+
+#include <dune/fem/function/adaptivefunction.hh>
+#include <dune/fem/function/vectorfunction.hh>
+#include <dune/fem/function/blockvectorfunction.hh>
+#include <dune/fem/function/blockvectordiscretefunction.hh>
+#include <dune/fem/space/discontinuousgalerkin.hh>
+#include <dune/fem/space/combinedspace.hh>
+#include <dune/fem/space/common/interpolate.hh>
+#include <dune/fem/io/file/dataoutput.hh>
+
+#include <dune/fem/quadrature/cachingquadrature.hh>
+#include <dune/fem/gridpart/adaptiveleafgridpart.hh>
+
+#include <dune/fem/misc/l2norm.hh>
+#include <dune/fem/misc/capabilities.hh>
+
+#include <dune/fem/io/parameter.hh>
+
+#include <dune/mmesh/mmesh.hh>
+// #include <dune/alugrid/grid.hh>
+
+using namespace Dune;
+using namespace Fem;
+
+typedef Dune::MovingMesh<2> MyGridType;
+// typedef Dune::ALUGrid<2, 2, simplex, conforming> MyGridType;
+typedef DGAdaptiveLeafGridPart< MyGridType > GridPartType;
+
+static const std::string usingSpaceName("Using TupleDiscreteFunctionSpace");
+typedef Dune::Fem::FunctionSpace< MyGridType::ctype, double, 2, 2 > FuncSpace1;
+typedef Dune::Fem::FunctionSpace< MyGridType::ctype, double, 2, 1 > FuncSpace2;
+typedef Dune::Fem::DiscontinuousGalerkinSpace< FuncSpace1, GridPartType, 2, CachingStorage > DiscreteFunctionSpaceType1;
+typedef Dune::Fem::DiscontinuousGalerkinSpace< FuncSpace2, GridPartType, 1, CachingStorage > DiscreteFunctionSpaceType2;
+typedef Dune::Fem::TupleDiscreteFunctionSpace< DiscreteFunctionSpaceType1, DiscreteFunctionSpaceType2 > DiscreteFunctionSpaceType;
+
+typedef typename DiscreteFunctionSpaceType :: FunctionSpaceType FunctionSpaceType;
+typedef AdaptiveDiscreteFunction< DiscreteFunctionSpaceType > DiscreteFunctionType;
+typedef DofManager< MyGridType > DofManagerType;
+typedef AdaptationManager< MyGridType, RestrictProlongDefault< DiscreteFunctionType > > AdaptationManagerType;
+
+// DataOutputParameters
+struct MyDataOutputParameters
+: public Dune::Fem::LocalParameter< DataOutputParameters, MyDataOutputParameters >
+{
+  MyDataOutputParameters ( const int step )
+  : step_( step )
+  {}
+
+  MyDataOutputParameters ( const MyDataOutputParameters &other )
+  : step_( other.step_ )
+  {}
+
+  std::string prefix () const
+  {
+    std::stringstream s;
+    s << "adapt-tuplespace-" << step_ << "-";
+    return s.str();
+  }
+
+private:
+  int step_;
+};
+
+// ***********************************************************
+// the exact solution to the problem for EOC calculation
+struct ExactSolution
+: public Fem::Function< FunctionSpaceType, ExactSolution >
+{
+  typedef FunctionSpaceType::RangeType RangeType;
+  typedef FunctionSpaceType::RangeFieldType RangeFieldType;
+  typedef FunctionSpaceType::DomainType DomainType;
+
+  //! f(x,y) = x*(1-x)*y*(1-y)
[…]
