### 2022-02-11 b6dd113
Use Lagrange and DG space in tuple space test.
