### 2022-02-11 #3959 @vijaysm
I had no success in getting the installation working on my Ubuntu (18.04) workstation either. I am unsure whether it was due to my CMake version not being up to the requirement. However, I'm going to try to see if I can get a clean environment on a Docker container with Ubuntu-latest to run your commands from the notebook to see if I can make progress. I apologize that this has taken more time than expected. I will document what I see and get back to you.

On a separate note, I would advise some form of simplification to the installation procedure. I understand that `pip install dune-mmesh` is as simple as it gets, but the installation has failed in more than one architecture. The errors seen should be better documented, and architecture-specific instructions if possible should be posted on the readthedocs page to help users get started easily.
