+++ doc/pyreference/trace.rst
@@ -2,6 +2,9 @@
 Trace and skeleton
 ******************
 
+Unfortunately, documentation of Python API for master can not be generated at the moment.
+Please rely on version [release/1.3.1](https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/trace.html) instead!
+
 .. automodule:: dune.mmesh
   :noindex:
   :members: trace, skeleton
