+++ doc/pyreference/solve.rst
@@ -2,6 +2,9 @@
 Coupled solve
 *************
 
+Unfortunately, documentation of Python API for master can not be generated at the moment.
+Please rely on version [release/1.3.1](https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/solve.html) instead!
+
 .. automodule:: dune.mmesh
   :noindex:
   :members: iterativeSolve, monolithicSolve
