+++ doc/pyreference/solve.rst
@@ -3,7 +3,7 @@ Coupled solve
 *************
 
 Unfortunately, documentation of Python API for master can not be generated at the moment.
-Please rely on version [release/1.3.1](https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/solve.html) instead!
+Please rely on version `release/1.3.1 <https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/solve.html>`_ instead!
 
 .. automodule:: dune.mmesh
   :noindex:
