+++ doc/contribute.rst
@@ -2,4 +2,5 @@
 Contribute
 **********
 
-If you want to contribute, you can do this using the `GitLab <https://gitlab.dune-project.org/samuel.burbulla/dune-mmesh>`_.
+If you want to contribute, you can do this using our `GitLab <https://gitlab.dune-project.org/samuel.burbulla/dune-mmesh>`_
+or `GitHub <https://github.com/samuelburbulla/dune-mmesh/>`_.
