+++ doc/pyreference/trace.rst
@@ -3,7 +3,7 @@ Trace and skeleton
 ******************
 
 Unfortunately, documentation of Python API for master can not be generated at the moment.
-Please rely on version [release/1.3.1](https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/trace.html) instead!
+Please rely on version `release/1.3.1 <https://dune-mmesh.readthedocs.io/en/release-1.3.1/pyreference/trace.html>`_ instead!
 
 .. automodule:: dune.mmesh
   :noindex:
