### 2022-02-15 22ebc2a
Fix version number for RTD documentation.
