### 2022-02-15 7537273
Add grid paramter to skeleton again in order to fix CI tests.
