### 2022-02-15 b3a5508
Improve distance computation and export it to python frontend.
