+++ dune/mmesh/CMakeLists.txt
@@ -10,5 +10,3 @@ set(HEADERS
 
 install(FILES ${HEADERS}
   DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/dune/mmesh)
-
-dune_add_library("dunemmesh" mmesh.cc)
