### 2022-02-22 0a40cc7
Use different mesh for distance test.
