### 2022-02-22 ecace50
Add test of distance in 3d.
