+++ dune/python/mmesh/test/distance.py
@@ -1,25 +1,36 @@
-# Grid creation
 from dune.grid import reader
 from dune.mmesh import mmesh
-dim = 2
-file = "horizontal.msh"
-gridView  = mmesh((reader.gmsh, file), dim)
-igridView = gridView.hierarchicalGrid.interfaceGrid
-
-from dune.mmesh import interfaceIndicator
-I = interfaceIndicator(igridView)
-
 from dune.mmesh import distance
-dist = distance(gridView)
-
 from ufl import grad
+from dune.fem.function import integrate
+
+
+# 2D
+file = "line2d.msh"
+grid = mmesh((reader.gmsh, file), 2)
+dist = distance(grid)
+
 normal = grad(dist)
-gridView.writeVTK("distance", pointdata={"d": dist, "n": [normal[0], normal[1], 0]}, nonconforming=True)
+grid.writeVTK("distance-2d", pointdata={"d": dist, "n": [normal[0], normal[1], 0]}, nonconforming=True)
+
+int = integrate(grid, dist, order=5)
+nor = integrate(grid, normal, order=5)
+
+assert(abs(int - 0.25) < 1e-8)
+assert((nor * nor) < 1e-8)
+
+
+# 3D
+file = "flat3d.msh"
+grid  = mmesh((reader.gmsh, file), 3)
+dist = distance(grid)
+
+normal = grad(dist)
+grid.writeVTK("distance", pointdata={"d": dist, "n": normal}, nonconforming=True)
 
 from dune.fem.function import integrate
-int = integrate(gridView, dist, order=5)
-jac = integrate(gridView, grad(dist), order=5)
+int = integrate(grid, dist, order=5)
+nor = integrate(grid, normal, order=5)
 
-print(int, jac)
-assert(abs(int - 0.273323) < 1e-6)
-assert(abs(jac[0]) < 1e-6 and abs(jac[1]) < 1e-6)
+assert(abs(int - 0.25) < 1e-3)
+assert((nor * nor) < 1e-8)
