+++ python/dune/mmesh/CMakeLists.txt
@@ -16,7 +16,8 @@ set(PRECOMPILE
   _interfaceindicator3d
   _normals2d
   _normals3d
-  _distance
+  _distance2d
+  _distance3d
 )
 
 foreach(obj ${PRECOMPILE})
