### 2022-02-22 146142e
Add pkg-config and OpenMPI to dependencies.
