+++ scripts/example.py
@@ -1,28 +1,4 @@
-# The Dune-MMesh Module
-
-The Dune-MMesh Module is an implementation of the DUNE grid interface that wraps CGAL triangulations in 2D and 3D. It is also capable to export a prescribed set of cell facets as a dim-1 interface grid and remesh the grid when moving this interface.
-
-You find the full documentation of Dune-MMesh at [dune-mmesh.readthedocs.io](https://dune-mmesh.readthedocs.io).
-
-
-## Installation
-
-We recommend the use a virtual environment. Simply install the Dune-MMesh package using pip
-````
-pip install dune-mmesh
-````
-
-More details on the installation procedure can be found on [Installation](https://dune-mmesh.readthedocs.io/en/latest/installation.html).
-
-
-## Example
-
-You can find a collection of examples of how to use Dune-MMesh on our [Examples](https://dune-mmesh.readthedocs.io/en/latest/examples.html) page.
-
-
-In a short example below we show how to use Dune-MMesh in general.
-We use `.msh` files generated by gmsh to define the geometry of our mesh.
-````
+# Generate mesh
 import gmsh
 gmsh.initialize()
 gmsh.option.setNumber("Mesh.MshFileVersion", 2.2)
@@ -32,7 +8,7 @@ gmsh.model.add(name)
 kernel = gmsh.model.occ
 box = kernel.addRectangle(0, 0, 0, 1, 1)
 p0 = kernel.addPoint(0.5, 0.5, 0, h)
-p1 = kernel.addPoint(0.25, 0.25, 0, h)
+p1 = kernel.addPoint(0.25, 0.5, 0, h)
 p2 = kernel.addPoint(0.75, 0.75, 0, h)
 p3 = kernel.addPoint(0.75, 0.25, 0, h)
 lf1 = kernel.addLine(p0, p1)
@@ -43,18 +19,16 @@ gmsh.model.mesh.embed(1, [lf1, lf2, lf3], 2, box)
 gmsh.model.mesh.generate(dim=2)
 gmsh.write(name)
 gmsh.finalize()
-````
 
-Read the generated `.msh` file and construct bulk and interface triangulation.
-````
+
+# Grid creation
 from dune.grid import reader
 from dune.mmesh import mmesh
 gridView  = mmesh((reader.gmsh, name), 2)
 igridView = gridView.hierarchicalGrid.interfaceGrid
-````
 
-Now, we can solve a mixed-dimensional PDE.
-````
+
+# Coupled problem
 from dune.mmesh import trace, skeleton, interfaceIndicator, monolithicSolve
 
 from dune.fem.space import dglagrange, lagrange
@@ -103,10 +77,8 @@ ischeme = galerkin([ia == ib])
 
 from dune.mmesh import monolithicSolve
 monolithicSolve(schemes=(scheme, ischeme), targets=(uh, iuh), verbose=True)
-````
 
-We can write the solution to `.vtk` or plot with matplotlib.
-````
+# Plot
 gridView.writeVTK("example", pointdata=[uh], nonconforming=True)
 igridView.writeVTK("interface", pointdata=[iuh])
 
@@ -116,21 +88,3 @@ fig = plt.figure()
 plot(uh, linewidth=0, clim=[0, 0.17], figure=fig)
 plot(iuh, linewidth=0.01, colorbar=None, clim=[0, 0.17], figure=fig)
 plt.savefig("plot.png")
-````
-
-![](scripts/plot.png)
-
-## Testing
-To run the Dune-MMesh tests, follow the installation instructions of the development version on the [Installation](https://dune-mmesh.readthedocs.io/en/latest/installation.html) page, and run
-````
-make build_tests
-make test
-````
-in the build directory.
-
-## Contribution
-
[…]
