+++ doc/installation.rst
@@ -13,6 +13,7 @@ In order to install and use Dune-MMesh you need:
 * OpenMPI (this will become optional)
 * Python (3.7 or later)
 
+
 There are two ways of installing Dune-MMesh.
 
 Using Pip
