### 2022-02-23 5b9b5c3
Move python tests and add 'python -m dune.mmesh test' command.
