+++ python/dune/mmesh/test/normals.py
@@ -5,7 +5,7 @@ from dune.fem.function import integrate
 
 
 # 2D
-file = "line2d.msh"
+file = "line.msh"
 grid = mmesh((reader.gmsh, file), 2)
 igrid = grid.hierarchicalGrid.interfaceGrid
 
@@ -19,7 +19,7 @@ assert(abs(int[1] - 1.) < 1e-8)
 
 
 # 3D
-file = "flat3d.msh"
+file = "plane.msh"
 grid  = mmesh((reader.gmsh, file), 3)
 igrid = grid.hierarchicalGrid.interfaceGrid
 
