+++ python/dune/mmesh/CMakeLists.txt
@@ -1,5 +1,8 @@
+add_subdirectory(test)
+
 add_python_targets(mmesh
   __init__
+  __main__
   _grids
   _move
   _skeletontrace
