### 2022-02-23 6ab90f9
Generate meshes with python and run all tests found.
