### 2022-02-23 f1fbd96
Update README entry about testing.
