### 2022-02-25 2a7e485
Add test.yml workflow for CI testing on GitHub.
