### 2022-02-25 8a5c594
Use new python -m dune.mmesh test command.
