+++ .github/workflows/testpypi.yml
@@ -60,18 +60,7 @@ jobs:
           pip install dune-mmesh
         shell: bash
 
-      - name: Setup dune-py
-        run: |
-          source dune-env/bin/activate
-          pip list
-          DUNE_CONTROL_PATH=dune-env setup-dunepy.py --opts=config.opts
-        shell: bash
-
       - name: Run test
         run: |
-          source dune-env/bin/activate
-          export DUNEPY_DISABLE_PLOTTING=1
-          cd scripts
-          python ../doc/examples/grids/horizontal.py
-          python test.py
+          python -m dune.mmesh test
         shell: bash
