+++ .gitlab-ci.yml
@@ -1,28 +1,0 @@
-before_script:
-  - source /duneci/bin/duneci-init-job
-  - duneci-install-module https://gitlab.dune-project.org/core/dune-common.git
-  - duneci-install-module https://gitlab.dune-project.org/core/dune-geometry.git
-  - duneci-install-module https://gitlab.dune-project.org/core/dune-grid.git
-  - duneci-install-module https://gitlab.dune-project.org/core/dune-localfunctions.git
-  - duneci-install-module https://gitlab.dune-project.org/core/dune-istl.git
-  - duneci-install-module https://gitlab.dune-project.org/extensions/dune-alugrid.git
-  - duneci-install-module https://gitlab.dune-project.org/dune-fem/dune-fem.git
-
-debian-11-gcc-9-20:
-  image: registry.dune-project.org/docker/ci/debian:11
-  script: duneci-standard-test
-  variables:
-    DUNECI_TOOLCHAIN: gcc-9-20
-    DUNECI_BRANCH: "master"
-    DUNECI_OPTS: "scripts/opts/ci-gcc.opts"
-    DUNECI_PARALLEL: 8
-
-ubuntu:20.04-clang-10-20:
-  image: registry.dune-project.org/docker/ci/ubuntu:20.04
-  script: duneci-standard-test
-  variables:
-    GIT_SSL_NO_VERIFY: 1
-    DUNECI_TOOLCHAIN: clang-10-20
-    DUNECI_BRANCH: "master"
-    DUNECI_OPTS: "scripts/opts/ci-gcc.opts"
-    DUNECI_PARALLEL: 8
