+++ .github/workflows/test.yml
@@ -18,7 +18,9 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
-
+        with:
+          lfs: true
+      
       - name: Install dependencies
         run: |
           if [ "$RUNNER_OS" == "Linux" ]; then
