### 2022-03-07 samuelburbulla/dune-mmesh#2 @krober10nd
I still am unable to install the package.

```
keithroberts@Keiths-MacBook-Pro-2 dune-mmesh % pip install dune-mmesh
DEPRECATION: Configuring installation scheme with distutils config files is deprecated and will no longer work in the near future. If you are using a Homebrew or Linuxbrew Python, please see discussion at https://github.com/Homebrew/homebrew-core/issues/76621
Collecting dune-mmesh
  Downloading dune-mmesh-1.3.1.post1.tar.gz (2.6 MB)
  Installing build dependencies ... -

\
error
  error: subprocess-exited-with-error

  × pip subprocess to install build dependencies did not run successfully.
  │ exit code: 1
  ╰─> [407 lines of output]
      Collecting setuptools
        Using cached setuptools-60.9.3-py3-none-any.whl (1.1 MB)
      Collecting wheel
        Using cached wheel-0.37.1-py2.py3-none-any.whl (35 kB)
      Collecting scikit-build
        Downloading scikit_build-0.13.1-py2.py3-none-any.whl (75 kB)
           ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 75.4/75.4 KB 2.2 MB/s eta 0:00:00
      Collecting cmake
        Downloading cmake-3.22.2-py2.py3-none-macosx_10_9_universal2.macosx_10_9_x86_64.macosx_11_0_arm64.macosx_11_0_universal2.whl (74.6 MB)
      Collecting ninja
        Using cached ninja-1.10.2.3-py2.py3-none-macosx_10_9_universal2.macosx_10_9_x86_64.macosx_11_0_arm64.macosx_11_0_universal2.whl (231 kB)
      Collecting requests
        Downloading requests-2.27.1-py2.py3-none-any.whl (63 kB)
           ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 63.1/63.1 KB 1.9 MB/s eta 0:00:00
      Collecting dune-fem<=2.8.0.6
        Downloading dune-fem-2.8.0.6.tar.gz (2.2 MB)
        Installing build dependencies: started
        Installing build dependencies: finished with status 'done'
        Getting requirements to build wheel: started
        Getting requirements to build wheel: finished with status 'done'
        Preparing metadata (pyproject.toml): started
        Preparing metadata (pyproject.toml): finished with status 'done'
      Collecting gmsh
        Using cached gmsh-4.8.4-py3-none-any.whl
      Collecting dune-grid<=2.8.0.6
        Using cached dune_grid-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting packaging
        Using cached packaging-21.3-py3-none-any.whl (40 kB)
      Collecting distro
        Using cached distro-1.7.0-py3-none-any.whl (20 kB)
      Collecting charset-normalizer~=2.0.0
        Using cached charset_normalizer-2.0.12-py3-none-any.whl (39 kB)
      Collecting certifi>=2017.4.17
        Using cached certifi-2021.10.8-py2.py3-none-any.whl (149 kB)
      Collecting urllib3<1.27,>=1.21.1
        Using cached urllib3-1.26.8-py2.py3-none-any.whl (138 kB)
      Collecting idna<4,>=2.5
        Using cached idna-3.3-py3-none-any.whl (61 kB)
      Collecting dune-alugrid<=v2.8.0.6
        Using cached dune_alugrid-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting dune-localfunctions<=v2.8.0.6
        Using cached dune_localfunctions-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting dune-istl<=v2.8.0.6
        Using cached dune_istl-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting fenics-ufl==2019.1.0
        Using cached fenics_ufl-2019.1.0-py3-none-any.whl (282 kB)
      Collecting numpy
        Using cached numpy-1.22.3-cp39-cp39-macosx_11_0_arm64.whl (12.8 MB)
      Collecting dune-geometry<=v2.8.0-r1
        Using cached dune_geometry-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting dune-common<=v2.8.0.post1
        Using cached dune_common-2.8.0.post1-cp39-cp39-macosx_12_0_arm64.whl
      Collecting pyparsing!=3.0.5,>=2.0.2
        Using cached pyparsing-3.0.7-py3-none-any.whl (98 kB)
      Collecting portalocker
        Using cached portalocker-2.4.0-py2.py3-none-any.whl (16 kB)
      Collecting mpi4py
        Using cached mpi4py-3.1.3-cp39-cp39-macosx_12_0_arm64.whl
      Building wheels for collected packages: dune-fem
        Building wheel for dune-fem (pyproject.toml): started
        Building wheel for dune-fem (pyproject.toml): finished with status 'error'
        error: subprocess-exited-with-error

        × Building wheel for dune-fem (pyproject.toml) did not run successfully.
        │ exit code: 1
        ╰─> [331 lines of output]
            /opt/homebrew/Cellar/python@3.9/3.9.10/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages/setuptools/dist.py:493: UserWarning: Normalizing 'v2.8.0.6' to '2.8.0.6'
              warnings.warn(tmpl.format(**locals()))
            Not searching for unused variables given on the command line.
            -- The C compiler identification is AppleClang 13.0.0.13000029
            -- Detecting C compiler ABI info
            -- Detecting C compiler ABI info - done
            -- Check for working C compiler: /Library/Developer/CommandLineTools/usr/bin/cc - skipped
            -- Detecting C compile features
            -- Detecting C compile features - done
            -- The CXX compiler identification is AppleClang 13.0.0.13000029
            -- Detecting CXX compiler ABI info
            -- Check for working CXX compiler: /Library/Developer/CommandLineTools/usr/bin/c++ - skipped
            -- Detecting CXX compile features
            -- Configuring done
            -- Generating done
            -- Build files have been written to: /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-eoa2zgm7/dune-fem_782549ab21f24dddb1a221cbfa245e01/_cmake_test_compile/build
            -- The C compiler identification is AppleClang 13.0.0.13000029
            -- The CXX compiler identification is AppleClang 13.0.0.13000029
            -- Detecting C compiler ABI info
            -- Detecting C compiler ABI info - done
            -- Check for working C compiler: /Library/Developer/CommandLineTools/usr/bin/cc - skipped
            -- Detecting C compile features
            -- Detecting C compile features - done
            -- Detecting CXX compiler ABI info
            -- Check for working CXX compiler: /Library/Developer/CommandLineTools/usr/bin/c++ - skipped
            -- Detecting CXX compile features
            -- Dependencies for dune-fem: dune-grid (>= 2.8)
            -- Suggestions for dune-fem: dune-alugrid (>= 2.8);dune-istl (>= 2.8);dune-localfunctions (>= 2.8);dune-spgrid (>= 2.8);dune-polygongrid
            -- Dependencies for dune-grid: dune-geometry (>= 2.8)
            -- Suggestions for dune-grid: dune-uggrid (>=2.8)
            -- Dependencies for dune-alugrid: dune-grid (>= 2.8)
            -- Dependencies for dune-istl: dune-common (>= 2.8)
            -- Dependencies for dune-localfunctions: dune-geometry (>= 2.8)
            -- Could NOT find dune-spgrid (missing: dune-spgrid_DIR)
            -- No full CMake package configuration support available. Falling back to pkg-config.
            -- Found PkgConfig: /opt/homebrew/bin/pkg-config (found version "0.29.2")
            -- Checking for module 'dune-spgrid>= 2.8'
            --   No package 'dune-spgrid' found
            -- Could NOT find dune-polygongrid (missing: dune-polygongrid_DIR)
            -- No full CMake package configuration support available. Falling back to pkg-config.
            -- Checking for module 'dune-polygongrid '
            --   No package 'dune-polygongrid' found
            -- Dependencies for dune-geometry: dune-common (>= 2.8)
            -- Dependencies for dune-geometry: dune-common (>= 2.8)
            -- Dependencies for dune-grid: dune-geometry (>= 2.8)
            -- Suggestions for dune-grid: dune-uggrid (>=2.8)
            -- Could NOT find dune-uggrid (missing: dune-uggrid_DIR)
            -- No full CMake package configuration support available. Falling back to pkg-config.
            -- Checking for module 'dune-uggrid>=2.8'
            --   No package 'dune-uggrid' found
            -- Could NOT find dune-uggrid (missing: dune-uggrid_DIR)
            -- No full CMake package configuration support available. Falling back to pkg-config.
            -- Checking for module 'dune-uggrid>=2.8'
            --   No package 'dune-uggrid' found
            -- Dependencies for dune-geometry: dune-common (>= 2.8)
            -- Performing Test cxx_std_flag_17
            -- Performing Test cxx_std_flag_17 - Success
            -- Performing Test compiler_supports_cxx17
            -- Performing Test compiler_supports_cxx17 - Success
            -- Performing Test HAS_ATTRIBUTE_UNUSED
            -- Performing Test HAS_ATTRIBUTE_UNUSED - Success
            -- Performing Test HAS_ATTRIBUTE_DEPRECATED
            -- Performing Test HAS_ATTRIBUTE_DEPRECATED - Success
            -- Performing Test HAS_ATTRIBUTE_DEPRECATED_MSG
            -- Performing Test HAS_ATTRIBUTE_DEPRECATED_MSG - Success
            -- Looking for std::experimental::make_array<int,int>
            -- Looking for std::experimental::make_array<int,int> - not found
            -- Looking for std::move<std::experimental::detected_t<std::decay_t,int>>
            -- Looking for std::move<std::experimental::detected_t<std::decay_t,int>> - found
            -- Looking for std::identity
            -- Looking for std::identity - not found
            -- Found LatexMk: /Library/TeX/texbin/latexmk (found version "Version 4.70b")
            -- Could NOT find Sphinx (missing: SPHINX_EXECUTABLE)
            -- Searching for macro file 'DuneCommonMacros' for module 'dune-common'.
            -- Performing tests specific to dune-common from file /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/DuneCommonMacros.cmake.
            -- Set Minimal Debug Level to 4
            -- Looking for sgemm_
            -- Looking for sgemm_ - not found
            -- Looking for pthread.h
            -- Looking for pthread.h - found
            -- Performing Test CMAKE_HAVE_LIBC_PTHREAD
            -- Performing Test CMAKE_HAVE_LIBC_PTHREAD - Success
            -- Found Threads: TRUE
            -- Looking for dgemm_
            -- Looking for dgemm_ - found
            -- Found BLAS: /Library/Developer/CommandLineTools/SDKs/MacOSX12.1.sdk/System/Library/Frameworks/Accelerate.framework
            -- Looking for cheev_
            -- Looking for cheev_ - found
            -- Found LAPACK: /Library/Developer/CommandLineTools/SDKs/MacOSX12.1.sdk/System/Library/Frameworks/Accelerate.framework;-lm;-ldl
            -- Looking for dsyev_
            -- Looking for dsyev_ - found
            -- Found GMP: /opt/homebrew/Cellar/gmp/6.2.1_1/lib/libgmpxx.dylib
            -- Performing Test QuadMath_COMPILES
            -- Performing Test QuadMath_COMPILES - Failed
            -- Could NOT find QuadMath (missing: QuadMath_VAR QuadMath_COMPILES)
            -- Found MPI_C: /opt/homebrew/Cellar/open-mpi/4.1.2/lib/libmpi.dylib (found suitable version "3.1", minimum required is "3.0")
            -- Found MPI: TRUE (found suitable version "3.1", minimum required is "3.0") found components: C
            -- Could NOT find TBB (set TBB_DIR to path containing TBBConfig.cmake or set PKG_CONFIG_PATH to include the location of the tbb.pc file) (missing: PkgConfigTBB_LINK_LIBRARIES PkgConfigTBB_FOUND) (found version "")
            -- Could NOT find PTScotch (missing: SCOTCH_LIBRARY SCOTCHERR_LIBRARY SCOTCH_INCLUDE_DIR)
            -- Could NOT find METIS (missing: METIS_LIBRARY METIS_INCLUDE_DIR METIS_API_VERSION)
            -- Could NOT find METIS (missing: METIS_LIBRARY METIS_INCLUDE_DIR METIS_API_VERSION)
            -- Found MPI_C: /opt/homebrew/Cellar/open-mpi/4.1.2/lib/libmpi.dylib (found version "3.1")
            -- Found MPI: TRUE (found version "3.1") found components: C
            -- Could NOT find ParMETIS (missing: PARMETIS_LIBRARY PARMETIS_INCLUDE_DIR METIS_FOUND) (Required is at least version "4.0")
            -- Could NOT find Vc (missing: Vc_DIR)
            -- Found Python3: /opt/homebrew/Frameworks/Python.framework/Versions/3.9/bin/python3 (found version "3.9.10") found components: Interpreter Development Development.Module Development.Embed
            -- Found pip_/opt/homebrew/Frameworks/Python.framework/Versions/3.9/bin/python3: TRUE
            -- Setting dune-common_INCLUDE_DIRS=/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/include
            -- Setting dune-common_LIBRARIES=dunecommon
            -- Searching for macro file 'DuneGeometryMacros' for module 'dune-geometry'.
            -- No module specific tests performed for module 'dune-geometry' because macro file 'DuneGeometryMacros.cmake' not found in /opt/homebrew/Cellar/python@3.9/3.9.10/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages/skbuild/resources/cmake;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-eoa2zgm7/dune-fem_782549ab21f24dddb1a221cbfa245e01/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules.
            -- Setting dune-geometry_INCLUDE_DIRS=/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/include
            -- Setting dune-geometry_LIBRARIES=dunegeometry
            -- Searching for macro file 'DuneUggridMacros' for module 'dune-uggrid'.
            -- No module specific tests performed for module 'dune-uggrid' because macro file 'DuneUggridMacros.cmake' not found in /opt/homebrew/Cellar/python@3.9/3.9.10/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages/skbuild/resources/cmake;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-install-eoa2zgm7/dune-fem_782549ab21f24dddb1a221cbfa245e01/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules;/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules.
            -- Searching for macro file 'DuneGridMacros' for module 'dune-grid'.
            -- Performing tests specific to dune-grid from file /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/DuneGridMacros.cmake.
            -- Looking for mkstemp
            -- Looking for mkstemp - found
            -- Could NOT find METIS (missing: METIS_LIBRARY METIS_INCLUDE_DIR METIS_API_VERSION)
            -- Could NOT find METIS (missing: METIS_LIBRARY METIS_INCLUDE_DIR METIS_API_VERSION)
            -- Could NOT find ParMETIS (missing: PARMETIS_LIBRARY PARMETIS_INCLUDE_DIR METIS_FOUND)
            -- Checking for module 'alberta-grid_1d>=3.0'
            --   No package 'alberta-grid_1d' found
            -- Checking for module 'alberta-grid_2d>=3.0'
            --   No package 'alberta-grid_2d' found
            -- Checking for module 'alberta-grid_3d>=3.0'
            --   No package 'alberta-grid_3d' found
            -- Checking for module 'alberta-grid_4d>=3.0'
            --   No package 'alberta-grid_4d' found
            -- Checking for module 'alberta-grid_5d>=3.0'
            --   No package 'alberta-grid_5d' found
            -- Checking for module 'alberta-grid_6d>=3.0'
            --   No package 'alberta-grid_6d' found
            -- Checking for module 'alberta-grid_7d>=3.0'
            --   No package 'alberta-grid_7d' found
            -- Checking for module 'alberta-grid_8d>=3.0'
            --   No package 'alberta-grid_8d' found
            -- Checking for module 'alberta-grid_9d>=3.0'
            --   No package 'alberta-grid_9d' found
            -- Could NOT find Alberta (set PKG_CONFIG_PATH to include the location of the alberta-grid_[n]d.pc files) (missing: ALBERTA_GRID_PREFIX) (Required is at least version "3.0")
            CMake Warning (dev) at /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/lib/python3.9/site-packages/cmake/data/CMake.app/Contents/share/cmake-3.22/Modules/FindPackageHandleStandardArgs.cmake:438 (message):
              The package name passed to `find_package_handle_standard_args` (PkgConfig)
              does not match the name of the calling package (Psurface).  This can lead
              to problems in calling code that expects `find_package` result variables
              (e.g., `_FOUND`) to follow a certain pattern.
            Call Stack (most recent call first):
              /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/lib/python3.9/site-packages/cmake/data/CMake.app/Contents/share/cmake-3.22/Modules/FindPkgConfig.cmake:99 (find_package_handle_standard_args)
              /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/FindPsurface.cmake:63 (include)
              /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/DuneGridMacros.cmake:20 (find_package)
              /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/DuneMacros.cmake:609 (include)
              /private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/share/dune/cmake/modules/DuneMacros.cmake:699 (dune_process_dependency_macros)
              CMakeLists.txt:25 (dune_project)
            This warning is for project developers.  Use -Wno-dev to suppress it.

            -- Checking for one of the modules 'psurface'
            -- Checking for module 'psurface'
            --   No package 'psurface' found
            -- Could NOT find Psurface (missing: PSURFACE_INCLUDE_DIR PSURFACE_LIBRARY)
            -- Could NOT find AmiraMesh (missing: AMIRAMESH_INCLUDE_DIR AMIRAMESH_LIBRARY)
            -- Setting dune-grid_INCLUDE_DIRS=/private/var/folders/9d/7ww_7jzx20j7kbrptmbv3kf00000gn/T/pip-build-env-q0ux788h/overlay/include
            -- Setting dune-grid_LIBRARIES=dunegrid
            -- Searching for macro file 'DunePolygongridMacros' for module 'dune-polygongrid'.
[…]
