### 2022-03-08 #3959 @vijaysm
It took me couple of iterations, but I think I have dune-mmesh successfully installed now in a Docker container. It may be simple to document this explicitly in your README for someone wanting to get started.
```
Steps to install Dune-MMesh on Ubuntu 20.04

1) docker run -it ubuntu:latest
2) apt update
3) apt install g++ cmake python3 python3-pip libboost-dev libopenmpi3 git
4) pip install dune-mmesh
```
I can now clone the repository and play with the examples locally. I have the image with dune-mmesh installed but have not had the time to experiment with the scripts. Meanwhile, any updates to the documentation for better procedures to install the library will be helpful, especially given the number of underlying dependencies.
