### 2022-03-08 #3959 @krober10nd
Okay. I've tried the identical set of steps and it's failed on my end although I don't have Ubuntu to test on. The fact that it's been this challenging to install is not a good sign. How are other users using this code? Can this be simplified?
