### 2022-03-09 #3959 @diehlpk
@samuelburbulla What about proving some Docker image with the latest stable version or master for testing the code?
