### 2022-03-15 ac61957
[bugfix] Fix interpolation from incident interface element.
