### 2022-03-16 #3959 @samuelburbulla
@vijaysm Thanks a lot for trying again and posting your Docker setup steps.

High appreciation for all of your feedback and your suggestions!

I agree that the installation is not yet clear enough, in particular regarding the large list of dependencies. So far, essentially people were using Dune-MMesh who had a working setup of DUNE on their system already, or a system which was preconfigured in that manner.

Reducing the requirements would be favourable, but is not easy, as many of them are implicit requirements of DUNE or dune-fem. I will try to make more of the dependencies optional (e.g. MPI is optional on the master already), but many of them will remain and are not installable via pip.

Following the setup of @vijaysm I added a more verbose description to the README.md and to the installation page. I think this makes things more clear and provides a simpler entry point. Can you have a look?

@diehlpk I would like to avoid providing a Docker image that has to be updated in the future. Do you think the updated documentation (including the docker setup) is sufficient?

@krobner10nd Have you tried to use the Docker setup with the ubuntu image? This is working on my Mac. And can you try following the steps I added for an installation on MacOS? This is how I use it on my Mac.
