### 2022-03-16 #3959 @diehlpk
> @diehlpk I would like to avoid providing a Docker image that has to be updated in the future. Do you think the updated documentation (including the docker setup) is sufficient?

I think a Docker image is not necessary, however, a Docker file like posted by @vijaysm and some instructions how to run it would be good.
