### 2022-03-16 79a1b7b
Small fixes in installation docs.
