### 2022-03-17 #3959 @samuelburbulla
Update: There is now a Dockerfile which builds a running container with preconfigured Dune-MMesh.
````
docker build -t mmesh \
  https://raw.githubusercontent.com/samuelburbulla/dune-mmesh/master/scripts/Dockerfile
docker run -it mmesh
````
This also works now on the Mac with its ARM architecture.
