### 2022-03-18 #3959 @diehlpk
> Update: There is now a Dockerfile which builds a running container with preconfigured Dune-MMesh.
>
> ```
> docker build -t mmesh \
>   https://raw.githubusercontent.com/samuelburbulla/dune-mmesh/master/scripts/Dockerfile
> docker run -it mmesh
> ```
>
> This also works now on the Mac with its ARM architecture.

@samuelburbulla That sounds good. You might want to add that to the README.md and documentation. Maybe you can write that if one just wants to test and play with your code, the Docker approach is the easiest way. Just put the commands how to run the docker image and possibly execute one example.

After that you can write that people, who want to implement new features or run heavier examples, should install the code and provide the install instructions.
