### 2022-04-06 1c5ee0e
Use shared_ptr in factories.
