+++ python/dune/mmesh/_grid3d.cc
@@ -10,7 +10,7 @@
 
 PYBIND11_MODULE( _grid3d, module )
 {
-  auto cls = Dune::Python::insertClass< Dune::MovingMesh<3> >( module,
+  auto cls = Dune::Python::insertClass< Dune::MovingMesh<3>, std::shared_ptr<Dune::MovingMesh<3>> >( module,
     "HierarchicalGrid",
     pybind11::dynamic_attr(),
     Dune::Python::GenerateTypeName("Dune::MovingMesh<3>"),
