### 2022-04-06 9c87e3f
Add addInterface to adaptation test.
