### 2022-04-06 9688f52
Initialize new interface elements in adaptation test.
