### 2022-04-06 1a0616d
Fix unsigned int warning and improve test.
