### 2022-04-12 44d5655
Specify package version requirements in dune.module.
