+++ setup.py
@@ -1,26 +1,13 @@
+#
+# DO NOT MODIFY THIS FILE!
+# This file is autogenerated by the `dunepackaging.py` script and
+# only used for the pypi dune packages. This file will not be included in
+# the build directory.
+#
+import os, sys
 try:
     from dune.packagemetadata import metaData
 except ImportError:
     from packagemetadata import metaData
 from skbuild import setup
-
-mmeshVersion = '1.4.dev0'
-duneFemVersion  = '2.8.0.6'
-
-metadata = metaData(duneFemVersion)[1]
-metadata['version'] = mmeshVersion
-
-# auto-generate pyproject.toml with duneVersion when building sdist
-from skbuild.command.sdist import sdist
-class mysdist(sdist):
-    def run(self):
-        requires = ['setuptools', 'wheel', 'scikit-build', 'cmake', 'ninja', 'requests']
-        requires += metadata['install_requires']
-        with open('pyproject.toml', 'w') as f:
-            f.write("[build-system]\n")
-            f.write("requires = ['"+"', '".join(requires)+"']\n")
-            f.write("build-backend = 'setuptools.build_meta'\n")
-        sdist.run(self)
-metadata['cmdclass'] = {'sdist': mysdist}
-
-setup(**metadata)
+setup(**metaData('1.4.0.dev20220412')[1])
