### 2022-04-21 #3959 @krober10nd
@diehlpk sure I can try this weekend, although I have left several more general issues and thoughts to consider on the github page for this package and paper that have yet to be addressed.
