### 2022-04-21 samuelburbulla/dune-mmesh#2 @samuelburbulla
Sorry, I have missed your last post... Thanks for posting the full log!

Obviously, the issue is a missing
````
/Library/Developer/CommandLineTools/SDKs/MacOSX12.0.sdk/usr/lib/libz.tbd
````
or a wrong search path for it.

Have you checked if this file is present?

On my Mac, it is located in a directory with a different version number of the MacOSX SDK.

It's hard to tell where the search path came from (it is deep in some CMake magic) and actually shouldn't go wrong, it should basically rely on CMake's find_package. I assume this is either an issue related to the M1, or it's just that the system updated in the meantime and some caches are still present.

A workaround could be to create a symlink of `MacOSX12.x.sdk` to `MacOSX12.0.sdk`, I have no better idea at the moment.
