+++ .github/workflows/test.yml
@@ -1,10 +1,6 @@
 name: test
 
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
+on: push
 
 env:
   BUILD_TYPE: Release
