### 2022-04-22 36fbdb1
Remove CMAKE_METADATA_FILE that is ignored and gives a CMake warning.
