+++ python/CMakeLists.txt
@@ -2,7 +2,6 @@ add_subdirectory(dune)
 
 dune_python_install_package(
   PATH "."
-  CMAKE_METADATA_FILE dune/mmesh/metadata.cmake
   DEPENDS _grid2d _grid3d _interfacegrid2d _interfacegrid3d
   CMAKE_METADATA_FLAGS DUNE_OPTS_FILE
 )
