### 2022-04-30 samuelburbulla/dune-mmesh#2 @krober10nd
Okay thanks, I was finally able to install but I wouldn't recommend anyone to try this on a Mac. There were a number of additional problems with Apple clang clashing with brew installed gcc.
