### 2022-05-12 e495486
Add L2 error measurement.
