+++ doc/examples/moving.py
@@ -5,12 +5,12 @@
 #
 # This is an example of how to move the interface and adapt the mesh.
 
-# We implement the finite volume moving mesh method presented in [CMR+18]_.
-#
+# We implement the finite volume moving mesh method presented in [CMR+18]_. 
+# 
 # .. [CMR+18] C. Chalons, J. Magiera, C. Rohde, M. Wiebe. A Finite-Volume Tracking Scheme for Two-Phase Compressible Flow. Theory, Numerics and Applications of Hyperbolic Problems I, pp. 309--322, 2018.
 
 # ## Grid creation
-#
+# 
 # We use the [vertical](grids/vertical.rst) grid file that contains an interface $\Gamma = {0.5} \times [0, 1]$ embedded in a domain $\Omega = [0,1]^2$.
 # For this example, we have to construct an _adaptive_ leaf grid view and we will need to obtain the hierarchical grid object.
 
@@ -26,8 +26,9 @@ gridView = adaptive( mmesh((reader.gmsh, file), dim) )
 hgrid = gridView.hierarchicalGrid
 igridView = hgrid.interfaceGrid
 
+
 # ## Problem
-#
+# 
 # Let us consider the following transport problem.
 # \begin{align}
 # u_t + \operatorname{div} f(u) = 0, & \qquad \text{in } \Omega \times [0,T], \\
@@ -38,7 +39,7 @@ igridView = hgrid.interfaceGrid
 # f(u) &= [1,0]^T u, \\
 # u_0(x,y) &= (0.5+x) \chi_{x<0.5}.
 # \end{align}
-#
+# 
 # Further, the interface is supposed to move with the transport speed in $f$, i.e. $m = [1,0]^T$.
 # \begin{align*}
 # \renewcommand{\jump}[1]{[\mskip-5mu[ #1 ]\mskip-5mu]}
@@ -69,13 +70,13 @@ def uexact(x, t):
 
 
 # ## Finite Volume Moving Mesh Method
-#
+# 
 # We use a Finite Volume Moving Mesh method to keep the discontinuity sharp. It can be formulated by
 # \begin{align}
 # \int_\Omega (u^{n+1} |det(\Psi)| - u^n) v\ dx + \Delta t \int_\mathcal{F} \big( g(u^{n}, n) - h(u^{n}, n) \big) \jump{v}\ dS = 0
 # \end{align}
 # where $\Psi := x + \Delta t s$ and s is a linear interpolation of the interface’s vertex movement m on the bulk triangulation.
-#
+# 
 # The numerical fluxes $g(u, n)$ and $h(u, n)$ are assumed to be consistent with the flux functions $f(u) \cdot n$ and $u s \cdot n$, respectively.
 
 # In[3]:
@@ -152,16 +153,22 @@ def dtCFL():
 
 
 # ## Timeloop without adaptation
-#
+# 
 # For comparison, we run the finite volume scheme once without adaptation of the mesh.
 
 # In[6]:
 
 
 from time import time
+from dune.fem.function import integrate
 from dune.fem.plotting import plotPointData as plot
 import matplotlib.pyplot as plt
-fig, axs = plt.subplots(1, 5, figsize=(12,3))
+plots = 4
+fig, axs = plt.subplots(1, plots, figsize=(3*plots,3))
+def shouldPlot(a):
+    t_a = a * tEnd / (plots-1)
+    return t.value > t_a and t.value - t_a <= tau.value
+
 runtime = 0
 
 uh.interpolate(u0(x))
@@ -185,14 +192,22 @@ while t.value < tEnd:
     runtime += time()
 
     i += 1
[…]
