+++ paper.md
@@ -29,15 +29,13 @@ bibliography: paper.bib
 
 # Summary
 
-Dune-MMesh is an implementation of the Dune [@BBD+21] grid interface that is tailored for numerical applications with possibly moving physical interfaces. The implementation based on CGAL triangulations [@CGAL] supports two and three dimensional meshes and can export a predefined set of facets as a separate interface grid. In spatial dimension two, arbitrary movement of vertices is enhanced with a re-meshing algorithm that implements non-hierarchical adaptation procedures. Various examples based on the python bindings of the discretization module dune-fem [@DNK20] have been implemented that demonstrate the versatile applicability of Dune-MMesh.
+In several physical and environmental processes that concern multiphase flows, biological systems, and geophysical phenomena, important physical processes occur along thin physical interfaces. These processes include effects that may alter the interface's position or topology over time creating a moving interface, which complicates traditional modeling techniques. Moving interface problems thus require advanced numerical tools with specific treatment of the interface and the simultaneous ability to implement complex physical effects, which this work seeks to create solutions for.
 
-# Statement of need
+# Statement of Need
 
-In many technical applications, in particular in the field of fluid dynamics, comparably thin physical interfaces can have a large impact on the overall behavior of a modeled system. Interfaces occur as separating layer between fluid phases in multiphase flows, in fluid-structure interaction, fluid-solid phase change and even fractures are modeled by lower-dimensional surfaces.
-
-The grid implementation Dune-MMesh aims at providing numerical capabilities for grid based methods to model interface-driven processes within the Dune framework. Essentially, it consists of two things: First, a triangulation based on CGAL where a set of facets is considered as interface. And, second, the possibility to re-mesh the triangulation when necessary.
-The representation of some grid facets as interface makes Dune-MMesh a useful tool for the implementation of mixed-dimensional models.
-The inevitable non-hierarchical adaptation complements the existing grid implementations within the Dune framework and allows for unprecedented flexibility of grid adaptation.
+In this work, we present Dune-MMesh that is tailored for numerical applications with moving physical interfaces. Dune-MMesh is an implementation of the well-developed Dune [@BBD+21] grid interface and is well-suited for the numerical discretization of partial differential equations. The package wraps two and three dimensional CGAL triangulations [@CGAL] in high-level objects like intersections of grid entities, index and id sets and geometry transformations and exports a predefined set of facets as a separate interface grid.
+In two dimensions, the arbitrary movement of vertices is enhanced with a re-meshing algorithm that implements non-hierarchical adaptation procedures. Besides the adaptation of the triangulation, Dune-MMesh provides the necessary data structures to adapt discrete functions defined on the bulk grid or the interface. This adaptation approach complements existing grid implementations within the Dune framework that strictly rely on hierarchical adaptation.
+Various examples in Python have been implemented based on the discretization module dune-fem [@DNK20] that demonstrate the versatile applicability of Dune-MMesh. Due to the ability to handle custom PDEs in their weak from written in Unified Form Language (UFL) and the mesh adaptation capabilities, we believe Dune-MMesh provides a useful tool for solving mixed-dimensional PDEs on moving interfaces that arise from various fields of modelling.
 
 # CGAL Wrapper
 
