### 2022-05-20 b9a52c1
Define some partition types.
