### 2022-05-23 #3959 @samuelburbulla
Thanks @vijaysm for your helpful comments! I will try to address your points below, and will make the changes right away on the feature/paper branch.

1. In the introduction section (L.8-L.12), you mention that both 2d and 3d meshes are supported using CGAL interface. However, you immediately talk about adaptivity or remeshing only in 2d with no mention of 3d. Is this intentional or an oversight? I noticed the same in L.51-L.52.

I must briefly mention that the introduction has changed somewhat due to comments from @krober10nd. However, the statement is still present and intended. The CGAL wrapper and the interface grid are fully implemented in 3d, but remeshing in 3d is not supported (yet). This is because removing vertices is not supported by the underlying CGAL Triangulation_3 class and more thought needs to be given to this.

2. In Figure 1, CGAL representations and DUNE numbering differ, and this is marked clearly. However, I only see face numbering for CGAL and no edge numbering in the 3D-Tet figure. I see that you mention in the text that CGAL doesn't provide one, but DUNE does? It would help to better clarify the legend and text.

Yes, DUNE does explicitly require a numbering of the edges. For instance, this can be relevant when defining finite-element spaces with Dofs at edges. I added a clarification to the article and extended the caption.

3. L.37: What do you mean by "Dune intersections" here? Are these referring to the interface intersection on a given mesh? Clarify or perhaps rephrase.

The term ‘Dune intersection’ means “Intersection of a mesh entity of codimension 0 with a neighboring element or with the domain boundary.” Clearly this needs to be clarified, and I added the definition to the article.

4. What are the light gray lines in the left figure of Figure 4? Is this a 2d or a 3d domain? It is confusing. Please clarify.

The gray and black lines each show a 2d triangulation. The gray triangulation is transformed by the blue vertex shifts into to black triangulation. I added a clarification to the caption of the figure.

5. L.56-57: To prevent degeneracy of cells, do you run mesh optimizations or smoothing procedures? Is this part of the remeshing routines? If yes, please expand accordingly. And if you do, are you using Dune or CGAL for it?

We apply no mesh optimizations or smoothing, just the remeshing procedures described thereafter in section “Adaptation”. As this is our own implementation, we use neither Dune nor CGAL. In L.58 I now refer explicitly to the adaptation section and within this section, I now elaborate on how we mark cells exactly.

6. L.64: "inavoidable" -> "unavoidable"

Corrected.

7. Suggestion: It might be worthwhile to show the objective error as a function of grid adaptation alongside the solution as well. Perhaps as a second row (subfigure). This will show how the error remains bounded or monotonically decreasing as the moving boundary problem is adaptively resolved.

I am not quite sure if I understand correctly what you suggest. In the moving example, I can plot the L2-Error between the numerical solution and the exact solution over time. This results in the following plot and shows that error increases over time as one would expect for such kind of time-dependent problems, I guess? So, where do you expect a bounded or monotonically decreasing error?

![Screenshot 2022-05-23 at 12 53 33](https://user-images.githubusercontent.com/34940349/169806107-41e44f2d-aba9-4f58-9158-fdb4d399d984.png)
