+++ dune/mmesh/test/examples/gmsh-implicit.cc
@@ -10,8 +10,9 @@
 #include <config.h>
 #include <dune/mmesh/mmesh.hh>
 
-int main()
+int main(int argc, char** argv)
 {
+  Dune::MPIHelper::instance(argc, argv);
   static constexpr int dim = GRIDDIM;
 
   using Grid = Dune::DelaunayTriangulation< dim >;
