+++ dune/mmesh/test/examples/dgf-interval.cc
@@ -4,8 +4,9 @@
 #include <config.h>
 #include <dune/mmesh/mmesh.hh>
 
-int main()
+int main(int argc, char** argv)
 {
+  Dune::MPIHelper::instance(argc, argv);
   static constexpr int dim = 2;
 
   using Grid = Dune::MovingMesh< dim >;
