### 2022-05-28 #3959 @vijaysm
I have been trying to run the examples but have not had luck so far. I have been able to compile and link the tests without any issues. However, running the tests does not yield any output whatsoever. For example, `test-grid-2d` and `test-femadaptation` under dune/mmesh/test just sit at 100% CPU usage even after 2 hours. Are these supposed to take this long?
