### 2022-05-30 e80e78d
Add workflow to build and push docker image to github packages.
