+++ .github/workflows/testpypi.yml
@@ -2,9 +2,9 @@ name: test pypi packages
 
 # Controls when the action will run.
 on:
-  # schedual test for nightly build
+  # schedule test for nightly build
   schedule:
-    - cron: "0 0 * * *"
+    - cron: "0 0 * * 0"
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
