### 2022-05-30 d271063
Create docker-image.yml
