### 2022-06-06 #3959 @vijaysm
@samuelburbulla So looks like the `.msh` files were just lfs references and so after I installed `git-lfs` as well, was able to git pull cleanly and get the full files. My further tries to run the examples still fail. I will open an issue in the repo related to this where we can further discuss and debug the issue.
