### 2022-06-06 #3959 @vijaysm
> I am not quite sure if I understand correctly what you suggest. In the moving example, I can plot the L2-Error between the numerical solution and the exact solution over time. This results in the following plot and shows that error increases over time as one would expect for such kind of time-dependent problems, I guess? So, where do you expect a bounded or monotonically decreasing error?

I meant that for any fixed point in time, adapting the mesh to the interface should produce a numerical solution that is much more accurate than the un-adapted one. Hence, my question was about plotting such an error convergence as a function of the adaptation levels/iterates, which I expect to decrease monotonically. The figure you shared above, in my understanding, is a time-dependent error, which has the temporal error component in addition to the spatial error due to moving interfaces. This certainly may not be monotone depending on how the temporal error propagates (immaterial of how well you adapt spatially).
