### 2022-06-10 #3959 @samuelburbulla
Of course, for any point in time we can compare the non-adapted solution to the adapted one, and the latter is always lower.
For the moving example this looks like this:

Probably I have to clarify one point: We do not adapt the mesh _to_ the interface, but actually move the corresponding interface edges. Therefore, there are no adaptation levels/iterates that reduce the error at the interface, it is already minimal. The adaptation of the mesh only happens next to the interface, but does not influence the discontinuity at the interface. Does that clear up the question?
