### 2022-06-15 #3959 @vijaysm
> Probably I have to clarify one point: We do not adapt the mesh to the interface, but actually move the corresponding interface edges. Therefore, there are no adaptation levels/iterates that reduce the error at the interface, it is already minimal. The adaptation of the mesh only happens next to the interface, but does not influence the discontinuity at the interface. Does that clear up the question?

Yes. That was exactly what I was asking. Thanks for clarifying this point. Perhaps this should also be mentioned in the paper as I had the impression that the grid adapts at the interface as well.
