### 2022-06-20 4686014
Set all unused entities to ghost.
