+++ dune/mmesh/test/test-l2projection.cc
@@ -1,0 +1,327 @@
+// include configurations options
+
+#ifdef HAVE_CONFIG_H
+#include <config.h>
+#endif
+
+// Includes from the IOStream Library
+// ----------------------------------
+
+#include <iostream>
+#include <sstream>
+
+//#define USE_OLD_RANNACHERTUREK_SPACE
+
+// Includes from DUNE-FEM
+// ----------------------
+
+// include Lagrange discrete function space
+#include <dune/fem/gridpart/adaptiveleafgridpart.hh>
+#include <dune/fem/space/lagrange.hh>
+#include <dune/fem/space/common/adaptationmanager.hh>
+#include <dune/fem/function/localfunction/bindable.hh>
+
+#if HAVE_PETSC && defined USE_PETSCDISCRETEFUNCTION
+#include <dune/fem/function/petscdiscretefunction.hh>
+#include <dune/fem/operator/linear/petscoperator.hh>
+#include <dune/fem/solver/petscinverseoperators.hh>
+#elif HAVE_DUNE_ISTL && defined USE_BLOCKVECTORFUNCTION
+#include <dune/fem/function/blockvectorfunction.hh>
+#include <dune/fem/operator/linear/istloperator.hh>
+#include <dune/fem/solver/istlinverseoperators.hh>
+#else
+#include <dune/fem/function/adaptivefunction.hh>
+#include <dune/fem/operator/linear/spoperator.hh>
+#if HAVE_SUITESPARSE_UMFPACK && defined USE_UMFPACK
+#include <dune/fem/solver/umfpacksolver.hh>
+#else
+#include <dune/fem/solver/krylovinverseoperators.hh>
+#endif
+#endif
+
+#include <dune/fem/space/common/interpolate.hh>
+
+#include <dune/fem/misc/l2norm.hh>
+#include <dune/fem/misc/h1norm.hh>
+
+#include <dune/fempy/function/virtualizedgridfunction.hh>
+
+#include "massoperator.hh"
+#include <dune/mmesh/mmesh.hh>
+
+// Global Type Definitions
+// -----------------------
+
+typedef Dune::MovingMesh<2> GridType;
+
+typedef Dune::Fem::AdaptiveLeafGridPart< GridType, Dune::InteriorBorder_Partition > GridPartType;
+typedef Dune::Fem::FunctionSpace< typename GridType::ctype, typename GridType::ctype, GridType::dimensionworld, 1 > SpaceType;
+typedef Dune::Fem::DynamicLagrangeDiscreteFunctionSpace< SpaceType, GridPartType > DiscreteSpaceType;
+//typedef Dune::Fem::LagrangeSpace< SpaceType, GridPartType > DiscreteSpaceType;
+
+#if HAVE_PETSC && defined USE_PETSCDISCRETEFUNCTION
+typedef Dune::Fem::PetscDiscreteFunction< DiscreteSpaceType > DiscreteFunctionType;
+typedef Dune::Fem::PetscLinearOperator< DiscreteFunctionType, DiscreteFunctionType > LinearOperatorType;
+typedef Dune::Fem::PetscInverseOperator< DiscreteFunctionType, LinearOperatorType > InverseOperatorType;
+#elif HAVE_DUNE_ISTL && defined USE_BLOCKVECTORFUNCTION
+typedef Dune::Fem::ISTLBlockVectorDiscreteFunction< DiscreteSpaceType > DiscreteFunctionType;
+typedef Dune::Fem::ISTLLinearOperator< DiscreteFunctionType, DiscreteFunctionType > LinearOperatorType;
[…]
