### 2022-06-20 #3959 @samuelburbulla
There is now a Docker image in the GitHub Container registry:
````
docker run -it ghcr.io/samuelburbulla/dune-mmesh:master
````

Thanks for clarifying, I added a sentence to the paper that we actually move the interface edges.
