### 2022-06-21 0fc5f27
Fix image path.
