### 2022-06-22 #3959 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.5281/zenodo.3706994 is OK
>
> MISSING DOIs
>
> - 10.1016/j.camwa.2020.06.007 may be a valid DOI for title: The DUNE framework: Basic concepts and recent developments
> - 10.1007/978-3-319-91545-6_25 may be a valid DOI for title: A Finite-Volume Tracking Scheme for Two-Phase Compressible Flow
>
> INVALID DOIs
>
> - None
> ```

@@samuel.burbulla please add the missing DOIs.
