### 2022-06-22 #3959 @diehlpk
From the editorial perspective please do the following changes:

*` University of Stuttgart 2 University of Warwick` -> please add the city and country

* `we will try to follow the general Dune approach and separate the adaptation into two stages.`

I assume that 1. Mark and 2. Adapt are the two stages. I would recommend to use Stage 1: Mark and Stage 2: Adapt to make it more clear.

After adding the missing DOIs and these two changes, we can proceed accepting the paper.
