### 2022-06-23 3d06678
Add dois, countries and stage.
