+++ paper.bib
@@ -5,7 +5,8 @@
   journal = {Computers \& Mathematics with Applications},
   volume = {81},
   year = {2021},
-  pages = {75--112}
+  pages = {75--112},
+  doi = {10.1016/j.camwa.2020.06.007}
 }
 
 @article{CGAL,
@@ -38,7 +39,8 @@
   title = {A Finite-Volume Tracking Scheme for Two-Phase Compressible Flow},
   journal = {Theory and Numerics and Applications of Hyperbolic Problems I},
   year = {2018},
-  pages = {309--322}
+  pages = {309--322},
+  doi = {10.1007/978-3-319-91545-6_25}
 }
 
 @article{GBK20,
