### 2022-06-23 #3959 @diehlpk
@samuelburbulla thanks for editing the paper.

Please do the following finalize the acceptance

- [ ] generate a new version, including all changes made during the review and post the new version here.
- [ ] upload this version to Zenodo or FigShare and post the DOI here. Note the title and the authors have to match with the psper.
