+++ python/dune/CMakeLists.txt
@@ -1,1 +1,5 @@
 add_subdirectory(mmesh)
+
+add_python_targets(dune
+  __init__
+)
