+++ pyproject.toml
@@ -1,14 +1,3 @@
-#
-# DO NOT MODIFY THIS FILE!
-# This file is autogenerated by the `dunepackaging.py` script and
-# only used for the pypi dune packages. This file will not be included in
-# the build directory.
-#
-# This is uses the `Python-Requires` field in the `dune.modules` file to
-# populate the `requires` entry. Additional packages needed for the package
-# build should be added in the `dune.modules`. These packages will then also be
-# included in the package install from source.
-#
 [build-system]
-requires = ['cmake>=3.13', 'dune-fem<=2.8.0.7', 'dune-grid<=2.8.0.post2', 'gmsh', 'ninja', 'pip', 'requests', 'scikit-build', 'setuptools', 'wheel']
+requires = ['setuptools', 'wheel', 'scikit-build', 'cmake>=3.13', 'ninja', 'requests', 'dune-fem<=2.8.0.7', 'gmsh', 'dune-grid<=2.8.0.post2']
 build-backend = 'setuptools.build_meta'
