+++ python/dune/mmesh/__init__.py
@@ -2,7 +2,7 @@ from ._grids import *
 from ._move import *
 from ._skeletontrace import *
 from ._solve import *
-from .utility import *
+from ._utility import *
 
 registry = dict()
 registry["grid"] = grid_registry
