+++ python/dune/mmesh/_utility.py
@@ -1,0 +1,228 @@
+import logging, traceback
+logger = logging.getLogger(__name__)
+
+import io
+from dune.generator import algorithm
+from dune.generator.exceptions import CompileError
+
+def interfaceIndicator(igrid, grid=None, restrict=True):
+    """Return indicator of interface edges.
+
+    Args:
+        igrid: The interfacegrid.
+        grid (Grid, optional): The bulk grid. Necessary, if wrapped.
+        restrict (bool, optional): If True, the returned UFL expression is restricted.
+
+    Returns:
+        Skeleton function that is one at interface edges.
+    """
+    from ufl import avg
+    from dune.mmesh import skeleton
+    try:
+        one = igrid.hierarchicalGrid.one
+    except:
+        from dune.fem.space import finiteVolume
+        space = finiteVolume(igrid)
+        one = space.interpolate(1, name="one")
+        igrid.hierarchicalGrid.one = one
+
+    if restrict:
+        return avg(skeleton(one, grid=grid))
+    else:
+        return skeleton(one, grid=grid)
+
+
+def domainMarker(grid):
+    """Return domain markers passed by .msh grid file.
+
+    Args:
+       grid: The grid.
+
+    Returns:
+       Grid function with cell-wise markers from .msh file.
+    """
+    code = """
+    #include <functional>
+    template <class GV>
+    auto domainMarker(const GV &gv) {
+      auto ret = [] (const auto& entity, const auto& xLocal) mutable -> auto {
+        return entity.impl().domainMarker();
+      };
+      return ret;
+    }
+    """
+    from dune.fem.space import finiteVolume
+    from dune.fem.function import cppFunction
+    try:
+        cppfunc = cppFunction(grid, name="domainMarker", order=0, fctName="domainMarker", includes=io.StringIO(code), args=[grid])
+    except CompileError as e:
+        code = code.replace("impl()", "impl().hostEntity().impl()")
+        cppfunc = cppFunction(grid, name="domainMarker", order=0, fctName="domainMarker", includes=io.StringIO(code), args=[grid])
+    fvspace = finiteVolume(grid)
+    return fvspace.interpolate(cppfunc, name="domainMarker")
+
+
+def interfaceDomainMarker(igrid):
+    """Return interface domain markers passed by .msh grid file.
+
+    Args:
+       igrid: The interface grid.
+
+    Returns:
+       Interface grid function with cell-wise markers from .msh file.
+    """
+    code = """
+    #include <iostream>
+    #include <functional>
+    template <class GV>
+    auto interfaceDomainMarker(const GV &igv) {
+      auto ret = [&igv] (const auto& entity, const auto& xLocal) mutable -> auto {
+        return igv.grid().domainMarker(entity);
+      };
+      return ret;
+    }
+    """
+    from dune.fem.function import cppFunction
+    return cppFunction(igrid, name="interfaceDomainMarker", order=0, fctName="interfaceDomainMarker", includes=io.StringIO(code), args=[igrid])
+
+
+def normals(igrid):
+    """Return normal vectors to the interface grid elements.
+
+    Args:
+        igrid: The interface grid.
+
+    Returns:
+        Grid function on the interface grid. Coincides with n('+') of the bulk facet normal.
+    """
+    code="""
+    #include <functional>
+    template <class IGV>
+    auto normal(const IGV &igv) {
+      return [&igv] (const auto& entity, const auto& xLocal) mutable -> auto {
+        return igv.grid().getMMesh().asIntersection( entity ).centerUnitOuterNormal();
+      };
+    }
+    """
+    import dune.ufl
+    from dune.fem.function import cppFunction
[…]

[…]
