### 2022-06-24 #3959 @samuelburbulla
I made a new release, v1.3.2, including all changes of the review and uploaded a new PyPI package.

The version is now published via Zenodo with DOI: 10.5281/zenodo.6705603
