### 2022-06-25 #3959 @Kevin-Mattheus-Moerman
- I checked the archive meta-data and all looks good.
- I checked the version tag for the review and archive and verified they match
- I proofread the paper and have the following comments:
- [ ] Should `dune-fem` have capital letters?
- [ ] `Use Partial Differential Equations (PDEs)` to introduce the acronym `PDEs`
- [ ] Check throughout for `Adaption`, this probably should read `Adaptation`
- [ ] Instead of `occur, e.g., in mixed-dimensional PDEs` I would recommend `occur, for example, in mixed-dimensional PDEs`

Please consider the above and update the paper here by calling `@editorialbot generate pdf`
