### 2022-06-27 #3959 @samuelburbulla
Thanks a lot @Kevin-Mattheus-Moerman for the detailed comments. I updated the paper accordingly.
